Refactoring: starting to extract db loading and reranking code from build_chain
