+++ src/reranking/__init__.py
@@ -1,0 +1,5 @@
+from .reranking_functions import compress_documents_lambda
+
+__all__ = [
+    "compress_documents_lambda",
+]
